Modernized a few bits of the assembler
* Added local labels to the idt_entry_handler and spinlock_xxx functions to reduce label clutter
* Refactored the spin lock code a bit. It no longer requires 4 bytes of stack space and the overal instruction count got reduced a bit

+++ sys/kern/spinlock.h
@@ -22,7 +22,7 @@
 #ifndef _SPINLOCK_H_
 #define _SPINLOCK_H_
 
-typedef uint8_t spinlock_t;
+typedef int32_t spinlock_t;
 
 #define SPINLOCK_INIT 0
 #define SPINLOCK_INIT_LOCKED 1
